% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mschart.R
\docType{data}
\name{mschart}
\alias{mschart}
\title{Microsoft Chart for .NET palettes}
\format{An object of class \code{list} of length 12.}
\usage{
mschart
}
\description{
This selection of the predifined color palettes in Microsoft Chart for .NET
version (3.5, 4.x).
}
\details{
The available palettes are:
Bright
Grayscale
Excel
Light
Pastel
EarthTones
SemiTransparent
Berry
Chocolate
Fire
SeaGreen
BrightPastel
}
\keyword{datasets}
