% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_palette.R
\name{palette_paste}
\alias{palette_paste}
\title{Converts a blob of text to a vector of colour palette}
\usage{
palette_paste(text)
}
\arguments{
\item{text}{A character vector of length 1 or more.}
}
\value{
A colour palette string stored in a vector form
}
\description{
Converts a blob of text to a vector of colour palette
}
