% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp_palettes.R
\docType{data}
\name{qp_art}
\alias{qp_art}
\title{Quickpalette artistic palettes}
\format{An object of class \code{list} of length 5.}
\usage{
qp_art
}
\description{
This selection of color palettes contains a list of 5 palettes with 10
colors each that act as being aesthetically pleasing.
}
\details{
The available palettes are:
dreaming
jojo
beach
waterfall
sunset
}
\keyword{datasets}
