% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickpalette-package.R
\docType{package}
\name{quickpalette-package}
\alias{quickpalette}
\alias{quickpalette-package}
\title{Quickly create palettes from outside sources}
\description{
Quickly create palettes from outside sources
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EmilHvitfeldt/quickpalette}
  \item Report bugs at \url{https://github.com/EmilHvitfeldt/quickpalette/issues}
}

}
\author{
\strong{Maintainer}: Emil Hvitfeldt \email{emilhhvitfeldt@gmail.com}

}
\keyword{internal}
