% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_palette.R
\name{regex_palette}
\alias{regex_palette}
\title{Extracts a color palette from a string}
\usage{
regex_palette(text)
}
\arguments{
\item{text}{A character vector of length 1 or more.}
}
\value{
A color palette as a string or a list of strings depending on the input.
}
\description{
This function takes rather messy strings and extracts the color palette
presented as hexadesimals and returns them as a string.
}
\examples{
text1 <- "3b59988b9dc3dfe3eef7f7f7ffffff"
text2 <- "the palatte is #3472bc, #345682 then #112233 and finally #cbac43"

# Returns vector when input is of length 1
regex_palette(text1)
regex_palette(text2)

text3 <- c("3b59988b9dc3dfe3eef7f7f7ffffff", "3b59988b9dc3dfe3eef7f7f7ffffff")

regex_palette(text3)
}
