% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_palette.R
\name{to_a_vector}
\alias{to_a_vector}
\title{Convert palette to a vector representation}
\usage{
to_a_vector(text)
}
\arguments{
\item{text}{a vector with colour codes}
}
\value{
A string of length 1 - vector representation that starts with c
}
\description{
Convert palette to a vector representation
}
