% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_palette.R
\name{vector_construct}
\alias{vector_construct}
\title{Construct a vector from either the clipboard or a text string}
\usage{
vector_construct(text)
}
\arguments{
\item{text}{A character vector of length 1 or more. Takes the most recent value from the clipboard if the input is missing}
}
\value{
A character vector - which is either the text from the clipboard or the input text
}
\description{
This function prepares the text for the regex_palette function
}
